Split selector logic into conda_build.selectors
